Fit : Ajout d'une verification du nombre de place + modification de kilometre en float

+++ ESP.Domain/Entities/Race.cs
@@ -11,7 +11,7 @@ public partial class Race
 
     public int IdRaceType { get; set; }
 
-    public double Kilometer { get; set; }
+    public float Kilometer { get; set; }
 
     public string Location { get; set; } = null!;
